requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.4
matplotlib>=3.8
seaborn>=0.13
openpyxl>=3.1
xgboost>=2.0
prophet>=1.1
torch>=2.2
shap>=0.45
scipy>=1.12
PyYAML>=6.0
